# Optimize `copy_cells`, part 2