--- a/tiledb/sm/global_state/global_state.h
+++ b/tiledb/sm/global_state/global_state.h
@@ -92,6 +92,13 @@ class GlobalState {
    */
   const std::string& cert_file();
 
+  /**
+   * Gets the number of threads that TBB was initialized with. Returns
+   * 0 if TBB is disabled or the global state has not been initialized.
+   * @return the number of configured TBB threads.
+   */
+  int tbb_threads();
+
  private:
   /** The TileDB configuration parameters. */
   Config config_;