… tiledb/sm/query/reader.cc